chore: consolidate l1 state mods
